Keep the latest year per traffic count
